Small type fix

+++ classes/Product.cs
@@ -2,10 +2,10 @@ class Product
 {
     public static int Id { get; set; }
     public static int Name { get; set; }
-    public static int Price { get; set; }
+    public static decimal Price { get; set; }
     public static int Quantity { get; set; }
 
-    public Product(int Id, int Name, int Price, int Quantity)
+    public Product(int Id, int Name, decimal Price, int Quantity)
     {
         Product.Id = Id;
         Product.Name = Name;
